Changed implementation to use multiprocessing instead of multithreading
